requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
python-dotenv==1.0.0
sentence-transformers==3.0.0
chromadb==1.3.4
groq==0.11.0
streamlit==1.38.0
pytest==8.3.0
httpx==0.27.0
